Add default ajax behavior to component
